requirements.txt
streamlit>=1.24.0
python-dotenv>=0.19.0
openai>=1.0.0
pygments==2.15.1
GitPython==3.1.37
requests>=2.31.0
diagrams>=0.23.3
graphviz==0.20.3
radon>=6.0.1
bandit>=1.8.3
safety>=3.5.1
mando>=0.7.1
colorama>=0.4.6
stevedore>=1.20.0
rich>=14.0.0
cryptography>=45.0.0
dparse>=0.6.4
filelock>=3.16.1
marshmallow>=4.0.0
nltk>=3.9.1
psutil>=6.1.0
pydantic>=2.9.2
ruamel-yaml>=0.18.10